Fix: Auth 공통 Style 컴포넌트
