+++ src/components/auth/styles/ButtonBox.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+import { Box } from '@mui/material';
+
+const ButtonBox = styled(Box)`
+	text-align: center;
+`;
+
+export default ButtonBox;
