+++ src/components/auth/styles/Form.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const Form = styled.form`
+	display: flex;
+	flex-direction: column;
+	gap: 30px;
+`;
+
+export default Form;
