+++ src/components/auth/styles/SubmitButton.tsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+import { Button } from '@mui/material';
+
+const SubmitButton = styled(Button)`
+	min-width: 150px;
+	height: 30px;
+	margin-top: 20px;
+`;
+
+export default SubmitButton;
