+++ src/components/auth/styles/index.ts
@@ -1,0 +1,3 @@
+export { default as SubmitButton } from './SubmitButton';
+export { default as Form } from './Form';
+export { default as ButtonBox } from './ButtonBox';
